fix: 엔티티 매핑 과정에서의 타입 불일치 해결 (#18)

+++ src/main/java/com/raisedeveloper/server/domain/user/domain/UserAlarmSettings.java
@@ -33,7 +33,7 @@ public class UserAlarmSettings extends CreatedUpdatedEntity {
 	private User user;
 
 	@Column(nullable = false)
-	private int alarmInterval;
+	private short alarmInterval;
 
 	@Column(nullable = false)
 	private LocalTime activeStartAt;
@@ -77,7 +77,7 @@ public class UserAlarmSettings extends CreatedUpdatedEntity {
 	}
 
 	public void updateSettings(
-		int interval,
+		short interval,
 		LocalTime activeStartAt,
 		LocalTime activeEndAt,
 		LocalTime focusStartAt,
